Generate MATLAB test assertions.
The mangling of longstr1 made this fail when we tried to use the
post-managled string accidentally; oops.

The strings are read in and the trailing newlines are kept, which
we have to trim when we print. And We have to output a gratuitous
trailing space for longstr2 because the parser insists on grabbing it
at position 105. Is there no trim() or strip() function?

Satrec is a monster with *100* fields!