Global satBillboards, only update Position in clock loop -- for speed.
Define the global satellite billboard collection as a global once.
Update positions only in the clock loop.
Split out the highlighting of selected satellites which touch the
global collection.

Created second collection for the GeoLocaiton but it causes a
satellite selection BUG whre we can't select with  the Earth disk #46