Push Cesium and worker down to own directory.
The workers *must* be started from their own file per AGI's SHunter,
and their regex wasn't finding my colocated file until I put it in its
own subdir with Cesium so it could be found.

The alternative was setting a global var for the URL path, or fixing
his regex.