Orbit traces rendered but coordinate frame wrong(?)
We have to add a PolylineCollection -- not just a Polyline -- to the
primitives to make it work.

We get a (meh) semi-smooth ellipse with 144 segments, but it can get
chunky.

More importantly, the orbit trace isn't where the satellite is. I
expect this is because the trace is not rotated into the same TEME
reference frame as the satellite billboard.

See #38