Display only one orbit trace at a time.
Thanks to Alex-developer for the pointer!

Make the orbitTraces a global polyline collection and remove all
polylines from it and add new when we display a new trace.

We could put any number of individual orbit traces (e.g., 10 different
satellite orbits) into the orbitTraces colleciton but only one is
stored at present. We'd have to determine a different way of
enabling/disabling individual traces, perhaps storing a list of traced
satellites, perhaps attached to the collection itself.

See #38