Use AGI's Cesium proxy to get ArcGIS map tiles.
AGI offers a /proxy/ service to a restricted set of servers including
the ArcGIS one that I'm using here. This commit is mainly to
demonstrate its use.  There are other maps, including roads, we could use.

We need our own for production, running on our server, to avoid CORS
restrictions accessing data.

(I was trying to get LANDSAT from NASA JPL OnEarth but that seems to
want a proprietary WMS request variant that I can't create -- and also
a proxy to a server which AGI doesn't allow.)