Get some missing TLEs from Celestrak search
Celestrak doesn't include TLEs in the .txt files for missions that are
no longer 'operational' (e.g., SMD says 'past'), even though they may
still be on orbit and findable by search (or y2no search/tracking). We
mark them as 'past' to signal nonoperational.

But we query for some of the missing ones anyway. need to review and
ensure the extra are really operational

23710 'RADARSAT-1'
26033 'ACRIMSAT'
22049 'GEOTAIL'